socialsync_ai/core/social_posting/utils.py
"""
social_posting/utils.py

Social media API connectors for SocialSync AI.
Currently uses mock print-based implementations.
Replace each function body with real API calls when credentials are ready.
"""

# ─────────────────────────────────────────────
# LinkedIn
# ─────────────────────────────────────────────

def post_to_linkedin(content: str) -> dict:
    """
    Post content to LinkedIn via the LinkedIn v2 API.

    TODO: Replace with real implementation:
        import requests
        headers = {"Authorization": f"Bearer {LINKEDIN_ACCESS_TOKEN}", ...}
        payload = { "author": f"urn:li:person:{PERSON_ID}", "lifecycleState": "PUBLISHED", ... }
        response = requests.post("https://api.linkedin.com/v2/ugcPosts", json=payload, headers=headers)
        response.raise_for_status()
        return response.json()
    """
    print(f"[LinkedIn Mock] 📤 Posting: {content[:80]}...")
    return {"status": "mock_success", "platform": "linkedin"}


# ─────────────────────────────────────────────
# Instagram
# ─────────────────────────────────────────────

def post_to_instagram(content: str) -> dict:
    """
    Post content to Instagram via the Instagram Graph API.

    TODO: Replace with real implementation:
        import requests
        # Step 1: Create media container
        # Step 2: Publish the container
        # Requires: INSTAGRAM_ACCOUNT_ID, INSTAGRAM_ACCESS_TOKEN
    """
    print(f"[Instagram Mock] 📤 Posting: {content[:80]}...")
    return {"status": "mock_success", "platform": "instagram"}


# ─────────────────────────────────────────────
# Twitter / X
# ─────────────────────────────────────────────

def post_to_twitter(content: str) -> dict:
    """
    Post content to Twitter/X via the Twitter API v2.

    TODO: Replace with real implementation:
        import tweepy
        client = tweepy.Client(
            consumer_key=TWITTER_API_KEY,
            consumer_secret=TWITTER_API_SECRET,
            access_token=TWITTER_ACCESS_TOKEN,
            access_token_secret=TWITTER_ACCESS_TOKEN_SECRET,
        )
        response = client.create_tweet(text=content)
        return response.data
    """
    print(f"[Twitter Mock] 📤 Posting: {content[:80]}...")
    return {"status": "mock_success", "platform": "twitter"}
